+++ default/linear_32/box.c
@@ -1,4 +1,4 @@
-/* $Id: box.c,v 1.2 2002/10/10 07:12:35 cegger Exp $
+/* $Id: box.c,v 1.3 2003/07/05 22:13:42 cegger Exp $
 ******************************************************************************
 
    Graphics library for GGI. Boxes.
@@ -72,12 +72,12 @@ int GGI_lin32_putbox(ggi_visual *vis, int x, int y, int w, int h, void *buffer)
 
 	/* Optimized full-width case */
 	if (w == destwidth && x == 0) {
-		memcpy(dest, src, w*h);
+		memcpy(dest, src, (size_t)(w*h));
 		return 0;
 	}
 	
 	while (h > 0) {
-		memcpy(dest, src, w);
+		memcpy(dest, src, (size_t)(w));
 		dest += destwidth;
 		src += srcwidth;
 		h--;
