+++ default/linear_32/hline.c
@@ -1,4 +1,4 @@
-/* $Id: hline.c,v 1.2 2002/10/10 07:12:35 cegger Exp $
+/* $Id: hline.c,v 1.3 2003/07/05 22:13:42 cegger Exp $
 ******************************************************************************
 
    Graphics library for GGI. Horizontal lines.
@@ -73,7 +73,7 @@ int GGI_lin32_puthline(ggi_visual *vis, int x, int y, int w, void *buffer)
 	PREPARE_FB(vis);
 
 	mem = (uint8 *)LIBGGI_CURWRITE(vis) + y*LIBGGI_FB_W_STRIDE(vis) + x*4;
-	memcpy(mem, buf32, w*4);
+	memcpy(mem, buf32, (size_t)(w*4));
 
 	return 0;
 }
@@ -87,7 +87,7 @@ int GGI_lin32_gethline(ggi_visual *vis, int x, int y, int w, void *buffer)
 
 	mem = (uint8 *)LIBGGI_CURREAD(vis) + y*LIBGGI_FB_R_STRIDE(vis) + x*4;
 
-	memcpy(buffer, mem, w*4);
+	memcpy(buffer, mem, (size_t)(w*4));
 
 	return 0;
 }
