+++ default/linear_4_r/hline.c
@@ -1,4 +1,4 @@
-/* $Id: hline.c,v 1.1 2001/05/12 23:01:45 cegger Exp $
+/* $Id: hline.c,v 1.2 2003/07/05 22:13:42 cegger Exp $
 ******************************************************************************
 
    Graphics library for GGI. Horizontal lines.
@@ -52,7 +52,7 @@ do_drawhline(ggi_visual *vis, int x, int y, int w)
 		w--;
 	}
 	
-	memset(fb, fg, w/2);
+	memset(fb, fg, (size_t)(w/2));
 
 	/* Dangling right pixel. */
 	if (w & 0x01) {
@@ -91,7 +91,7 @@ int GGI_lin4r_puthline(ggi_visual *vis, int x, int y, int w, void *buffer)
 	
 	/* Might as well take the memcpy speed up (multiple byte copy). */
 	if (!(x & 0x01)) {
-		memcpy(fb,buf,w/2);
+		memcpy(fb,buf, (size_t)(w/2));
 		if (w & 0x01)
 			*(fb+(w/2)) = ( *(fb+(w/2)) & 0x0F)
 				| (*((uint8 *)buf+(w/2)) << 4);
@@ -134,7 +134,7 @@ int GGI_lin4r_gethline(ggi_visual *vis,int x,int y,int w,void *buffer)
 	 * byte will hold an extraneous pixel.  No big deal. */
 	
 	if (!(x & 0x01)) {
-		memcpy(buf, fb, (w/2) + (w & 0x01));
+		memcpy(buf, fb, (size_t)((w/2) + (w & 0x01)));
 		return 0;
 	}
