+++ default/linear_8/hline.c
@@ -1,4 +1,4 @@
-/* $Id: hline.c,v 1.2 2002/10/09 22:41:40 cegger Exp $
+/* $Id: hline.c,v 1.3 2003/07/05 22:13:42 cegger Exp $
 ******************************************************************************
 
    Graphics library for GGI. Horizontal lines.
@@ -36,7 +36,7 @@ int GGI_lin8_drawhline(ggi_visual *vis,int x,int y,int w)
 	PREPARE_FB(vis);
 
 	memset((uint8*)LIBGGI_CURWRITE(vis) + y*LIBGGI_FB_W_STRIDE(vis) + x,
-	       LIBGGI_GC_FGCOLOR(vis), w);
+	       (signed)LIBGGI_GC_FGCOLOR(vis), (size_t)(w));
 
 	return 0;
 }
@@ -46,7 +46,7 @@ int GGI_lin8_drawhline_nc(ggi_visual *vis,int x,int y,int w)
 	PREPARE_FB(vis);
 
         memset((uint8*)LIBGGI_CURWRITE(vis) + y*LIBGGI_FB_W_STRIDE(vis) + x,
-	       LIBGGI_GC_FGCOLOR(vis), w);
+	       (signed)LIBGGI_GC_FGCOLOR(vis), (size_t)(w));
 
 	return 0;
 }
@@ -59,7 +59,7 @@ int GGI_lin8_puthline(ggi_visual *vis,int x,int y,int w,void *buffer)
 	PREPARE_FB(vis);
 
 	mem = (uint8*)LIBGGI_CURWRITE(vis) + y*LIBGGI_FB_W_STRIDE(vis) + x;
-	memcpy(mem, buf8, w);
+	memcpy(mem, buf8, (size_t)(w));
 
 	return 0;
 }
@@ -72,7 +72,7 @@ int GGI_lin8_gethline(ggi_visual *vis,int x,int y,int w,void *buffer)
 
 	mem = (uint8 *)LIBGGI_CURREAD(vis) + y*LIBGGI_FB_R_STRIDE(vis) + x;
 
-	memcpy(buffer, mem, w);
+	memcpy(buffer, mem, (size_t)(w));
 
 	return 0;
 }
